refactor: vertex drawers moved to Cairo or Matplotlib-specific submodules
